requirements.txt
# Python 3.14 compatible requirements
# TensorFlow is NOT supported on Python 3.14 — use PyTorch instead

torch>=2.10.0
torchvision>=0.21.0
torchaudio>=2.10.0

opencv-python>=4.10.0
numpy>=2.0.0
matplotlib>=3.9.0
scikit-learn>=1.5.0
pandas>=2.2.0
seaborn>=0.13.0
Pillow>=11.0.0
tqdm>=4.66.0
streamlit>=1.38.0
